Add files via upload
This project presents a secure and scalable network design for a university campus consisting of three buildings (100, 99, and 154). It focuses on efficient communication between departments using a structured three-layer architecture: core, distribution, and access. Each department is separated into VLANs to enhance control and security. Inter-VLAN communication is enabled through router-on-a-stick, while RIP is used as the routing protocol. ACLs are implemented to restrict access between VLANs. The network includes services like DHCP, DNS, web, and email servers, with SSH and Telnet configured for remote access. Redundant trunk links are added for reliability. The entire setup was simulated and tested using Cisco Packet Tracer, making it a practical model for real-world educational networks.